+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 v0.46.0
 -------
 * Increased default socket timeouts to 30s
-* Fixed market_map_data (the result has changed due to tsetmc API changes)
+* Fixed ``market_map_data`` (the result has changed due to tsetmc API changes)
 * Updated dataset
 
 v0.45.1