+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Unreleased
 ----------
 * Added ``market_watch.MarketWatch`` class that helps with watching the market.
 * Avoid unsigned data-types in returned values. This should help with some overflow bugs when user is not careful.
+* Updated dataset
 
 v0.46.0
 -------